Remove duplicate implementation
