+++ src/platform/packages/shared/kbn-spaces-utils/index.ts
@@ -8,4 +8,4 @@
  */
 
 export { DEFAULT_SPACE_ID } from './src/common/constants';
-export { getSpaceIdFromPath } from './src/common/spaces_url_parser';
+export { addSpaceIdToPath, getSpaceIdFromPath } from './src/common/spaces_url_parser';
