+++ src/platform/packages/shared/kbn-spaces-utils/src/common/spaces_url_parser.test.ts
@@ -7,8 +7,8 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
+import { addSpaceIdToPath, getSpaceIdFromPath } from './spaces_url_parser';
 import { DEFAULT_SPACE_ID } from './constants';
-import { getSpaceIdFromPath } from './spaces_url_parser';
 
 describe('getSpaceIdFromPath', () => {
   describe('without a serverBasePath defined', () => {
@@ -85,5 +85,27 @@ describe('getSpaceIdFromPath', () => {
         pathHasExplicitSpaceIdentifier: false,
       });
     });
+  });
+});
+
+describe('addSpaceIdToPath', () => {
+  test('handles no parameters', () => {
+    expect(addSpaceIdToPath()).toEqual(`/`);
+  });
+
+  test('it adds to the basePath correctly', () => {
+    expect(addSpaceIdToPath('/my/base/path', 'url-context')).toEqual('/my/base/path/s/url-context');
+  });
+
+  test('it appends the requested path to the end of the url context', () => {
+    expect(addSpaceIdToPath('/base', 'context', '/final/destination')).toEqual(
+      '/base/s/context/final/destination'
+    );
+  });
+
+  test('it throws an error when the requested path does not start with a slash', () => {
+    expect(() => {
+      addSpaceIdToPath('', '', 'foo');
+    }).toThrowErrorMatchingInlineSnapshot(`"path must start with a /"`);
   });
 });
