+++ x-pack/platform/plugins/shared/spaces/common/index.ts
@@ -13,7 +13,7 @@ export {
   DEFAULT_SPACE_ID,
   API_VERSIONS,
 } from './constants';
-export { addSpaceIdToPath, getSpaceIdFromPath } from './lib/spaces_url_parser';
+export { addSpaceIdToPath, getSpaceIdFromPath } from '@kbn/spaces-utils';
 export type {
   Space,
   GetAllSpacesOptions,
